requirements.txt
Flask
openai
gspread
oauth2client
python-dotenv
fastapi
uvicorn
jinja2
python-multipart